Move InnovationHubSettings to correct folder
