+++ src/domain/innovationHub/InnovationHubsAdmin/InnovationHubsAdmin.graphql
@@ -22,52 +22,15 @@ mutation deleteInnovationHub($innovationHubId: UUID!) {
   }
 }
 
-fragment InnovationHubProfile on Profile {
-  id
-  displayName
-  description
-  tagline
-  tagset {
-    ...TagsetDetails
-  }
-  visual(type: BANNER_WIDE) {
-    ...VisualFull
-  }
-  url
-}
-
-query InnovationHubSettings($innovationHubId: UUID!) {
-  platform {
-    id
-    innovationHub(id: $innovationHubId) {
-      ...AdminInnovationHub
-    }
-  }
-}
-
-fragment AdminInnovationHub on InnovationHub {
-  id
-  nameID
-  subdomain
-
-  profile {
-    ...InnovationHubProfile
-  }
-  spaceListFilter {
-    ...InnovationHubSpace
-  }
-
-  spaceVisibilityFilter
-}
 
 mutation createInnovationHub($hubData: CreateInnovationHubOnAccountInput!) {
   createInnovationHub(createData: $hubData) {
-    ...AdminInnovationHub
+    ...InnovationHubSettings
   }
 }
 
 mutation updateInnovationHub($hubData: UpdateInnovationHubInput!) {
   updateInnovationHub(updateData: $hubData) {
-    ...AdminInnovationHub
+    ...InnovationHubSettings
   }
 }
