+++ src/domain/innovationHub/InnovationHubsSettings/InnovationHubAvailableSpaces.graphql
