+++ src/domain/innovationHub/InnovationHubsSettings/InnovationHubSpace.graphql
